+++ src/refactoring/hooks/useProductManager.ts
@@ -1,0 +1,101 @@
+import { useState } from "react";
+import { Product, Discount } from "../../types";
+
+export const useProductManager = (
+  products: Product[],
+  onProductUpdate: (product: Product) => void
+) => {
+  const [openProductIds, setOpenProductIds] = useState<Set<string>>(new Set());
+  const [editingProduct, setEditingProduct] = useState<Product | null>(null);
+  const [newDiscount, setNewDiscount] = useState<Discount>({
+    quantity: 0,
+    rate: 0,
+  });
+  const [showNewProductForm, setShowNewProductForm] = useState(false);
+  const [newProduct, setNewProduct] = useState<Omit<Product, "id">>({
+    name: "",
+    price: 0,
+    stock: 0,
+    discounts: [],
+  });
+
+  const toggleProductAccordion = (productId: string) => {
+    setOpenProductIds((prev) => {
+      const newSet = new Set(prev);
+      newSet.has(productId) ? newSet.delete(productId) : newSet.add(productId);
+      return newSet;
+    });
+  };
+
+  const handleEditProduct = (product: Product) => {
+    setEditingProduct({ ...product });
+  };
+
+  const handleProductNameUpdate = (productId: string, newName: string) => {
+    if (editingProduct && editingProduct.id === productId) {
+      const updatedProduct = { ...editingProduct, name: newName };
+      setEditingProduct(updatedProduct);
+    }
+  };
+
+  const handlePriceUpdate = (productId: string, newPrice: number) => {
+    if (editingProduct && editingProduct.id === productId) {
+      setEditingProduct({ ...editingProduct, price: newPrice });
+    }
+  };
+
+  const handleStockUpdate = (productId: string, newStock: number) => {
+    if (editingProduct && editingProduct.id === productId) {
+      setEditingProduct({ ...editingProduct, stock: newStock });
+    }
+  };
+
+  const handleAddDiscount = (productId: string) => {
+    if (editingProduct && editingProduct.id === productId) {
+      const newProduct = {
+        ...editingProduct,
+        discounts: [...editingProduct.discounts, newDiscount],
+      };
+      onProductUpdate(newProduct);
+      setEditingProduct(newProduct);
+      setNewDiscount({ quantity: 0, rate: 0 });
+    }
+  };
+
+  const handleRemoveDiscount = (productId: string, index: number) => {
+    if (editingProduct && editingProduct.id === productId) {
+      const newProduct = {
+        ...editingProduct,
+        discounts: editingProduct.discounts.filter((_, i) => i !== index),
+      };
+      onProductUpdate(newProduct);
+      setEditingProduct(newProduct);
+    }
+  };
+
+  const handleEditComplete = () => {
+    if (editingProduct) {
+      onProductUpdate(editingProduct);
+      setEditingProduct(null);
+    }
+  };
+
+  return {
+    openProductIds,
+    editingProduct,
+    newDiscount,
+    showNewProductForm,
+    newProduct,
+    setNewDiscount,
+    setShowNewProductForm,
+    setNewProduct,
+    toggleProductAccordion,
+    handleEditProduct,
+    handleProductNameUpdate,
+    handlePriceUpdate,
+    handleStockUpdate,
+    handleAddDiscount,
+    handleRemoveDiscount,
+    handleEditComplete,
+  };
+};
